+++ source/EngineGpuKernels/TOs.cuh
@@ -53,7 +53,7 @@ struct ConstructorTO
 {
     // Properties
     uint8_t autoTriggerInterval;  // 0 = manual (triggered by signal), > 0 = auto trigger
-    uint32_t constructionActivationTime;
+    uint16_t constructionActivationTime;
 
     // Genome data
     uint16_t genomeSize;
